+++ lib/simple_https_service.dart
@@ -1,5 +1,6 @@
 export 'src/manager/timing_manager.dart';
 export 'src/network/enum_post_type.dart';
+export 'src/network/multipart_file_spec.dart';
 export 'src/network/util_check_url.dart';
 export 'src/network/https_service_stub.dart'
     if (dart.library.js_interop) 'src/network/https_service_web.dart';